move layout to layout/main, so later we can have layout/dev for dev resources
