+++ toktok/src/pages/blog.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link, graphql } from "gatsby";
 
 import Bio from "../components/bio";
-import Layout from "../components/layout";
+import Layout from "../components/layout/main";
 import SEO from "../components/seo";
 
 const BlogIndex = ({ data, location }) => {
