+++ toktok/src/pages/clients.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Layout from "../components/layout.tsx";
+import Layout from "../components/layout/main.tsx";
 
 const Page = () => (
   <Layout>
