+++ toktok/src/pages/download.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Layout from "../components/layout.tsx";
+import Layout from "../components/layout/main.tsx";
 import Link from "gatsby-link";
 
 const Page = () => (
