+++ toktok/src/pages/features.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import Layout from "../components/layout.tsx";
+import Layout from "../components/layout/main.tsx";
 import Features from "../components/content/features.tsx";
 
 const Page = () => (
