+++ toktok/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import Link from "gatsby-link";
-import Layout from "../components/layout.tsx";
+import Layout from "../components/layout/main.tsx";
 
 import About from "../components/content/about.tsx";
 import Features from "../components/content/features.tsx";
